social_media_platform/requirements.txt
authlib~=1.2
django~=4.2
python-dotenv~=1.0
requests~=2.31
djangorestframework
psycopg2-binary
django-cors-headers
python-jose
djangorestframework-simplejwt
